@@ -6,7 +6,7 @@ authors = [
     {name = "eric-castellanos",email = "ericcastellanos229@gmail.com"}
 ]
 readme = "README.md"
-requires-python = ">=3.9,<3.12"
+requires-python = ">=3.10,<3.12"
 dependencies = [
 ]
 
@@ -53,3 +53,13 @@ mlflow = "2.14.3"
 boto3 = "1.38.23"
 psycopg2-binary = "2.9.9"
 requests = "2.32.3"
+setuptools = "^75.0.0"
+
+[tool.poetry.group.model.dependencies]
+xgboost = { version = ">=3.1.2,<4.0.0", markers = "python_version >= '3.10' and python_version < '3.12'" }
+polars = "1.30.0"
+click = "8.1.8"
+scikit-learn = "^1.4.0"
+mlflow = "2.14.3"
+setuptools = "^75.0.0"
+optuna = "^3.6.1"
